seminar-03/opeani_research_agent/backend/requirements.txt
openai-agents==0.0.14
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
asyncio>=3.4.3
fastapi>=0.100.0
uvicorn>=0.22.0
sqlmodel>=0.0.8
alembic>=1.10.0
psycopg2-binary>=2.9.6
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5 